Smart import with module resolution, epoch 6, macOS CI (#685)
* Smart import with module resolution, epoch 6, macOS CI

Add three import primitives with ELLE_PATH-based resolution:
- import: resolves Elle modules (foo.elle, foo.lisp, libelle_foo.{so,dylib})
- import-native: resolves native plugins only (libelle_foo.{so,dylib})
- import-file: exact filesystem path (unchanged)

Resolution uses ELLE_HOME (auto-detected from binary path) and ELLE_PATH
(colon-separated search dirs, defaults to lib/ + target/{self,other}).
Relative imports (./foo) resolve from the importing file's directory via
a new source_file_stack on the VM.

Epoch 6 introduces the new primitives. elle rewrite migrates old
import-file calls: lib paths become (import "name"), plugin paths become
(import-native "name"). --dry-run now shows each edit.

Add macOS plugin builds and tests to CI (23 plugins × 2 platforms).

* Move cranelift deps out of Linux-only target section

Cranelift JIT dependencies were under [target.'cfg(target_os = "linux")']
alongside io-uring, causing compilation failure on macOS. Move them to
platform-independent [dependencies] where they belong.

* Fix macOS build: mode_t variadic promotion, cfg-gated unused vars

- Cast mode_t to c_uint for openat() variadic arg (mode_t is u16 on
  macOS, variadic C functions require at least c_uint)
- Prefix timeout and network_pool with underscore where they're only
  used inside #[cfg(target_os = "linux")] blocks

* Install coreutils on macOS CI for timeout command

GNU timeout (from coreutils) is used by the Makefile but isn't available
on macOS by default. brew install coreutils provides it as gtimeout and
adds it to PATH as timeout.

* Gate buffer_handle_mut behind cfg(linux), only used by io_uring

* Fix connect_fd panic on macOS, revert macOS plugin CI

The ThreadPool drain path in drain_platform_completions didn't stash
the connect fd before calling process_raw_completion, causing a panic
on macOS where all connects go through the thread pool (no io_uring).
Also fix >= 0 check (fd 0 is valid).

Drop macOS plugin builds/tests from CI — no faster than Linux. Keep
macOS smoke tests (build + examples + scripts).

* Fix subprocess/wait treating signal kills as waitid errors

The ProcessWait completion handler treated all negative result_code as
waitid syscall failures. On the thread pool path (macOS, and Linux
fallback), negative result_code means "killed by signal" (e.g. -15 for
SIGTERM from waitpid's WTERMSIG). Move the error check inside the
io_uring branch where negative truly means syscall failure.

* Increase CI timeout to 60s (macOS runners are slower)

* Poll-then-accept in thread pool to fix macOS exit hang

Replace blocking accept() with poll(500ms)+accept loop. When the
listener fd is closed, poll returns POLLNVAL or fstat returns EBADF,
and the worker thread exits cleanly. Without this, accept() blocks
indefinitely on macOS even after the fd is closed from another thread
(POSIX leaves this undefined), preventing process exit.

Revert the 60s CI timeout — 30s is sufficient now that the hang is fixed.

* Call process::exit on all paths to avoid macOS thread hang

On macOS (no io_uring), thread pool worker threads blocked in syscalls
can outlive the scheduler. The orderly shutdown (pump drain, fiber
completion, I/O cancellation) happens before this point; process::exit
is a safety net for any straggler OS threads. The error path already
called process::exit(1); now the success path does too.

* Fix thread pool port/close: generate error completions for cancelled ops

port/close on the ThreadPool path removed pending entries from the
backend's map but didn't generate error completions. The scheduler's
pump loop still had the op in its own pending map and blocked forever
in io/wait(-1). On io_uring, IORING_OP_ASYNC_CANCEL produces a CQE
that unblocks the scheduler; the thread pool path now does the same
by pushing an error completion.

Also add ELLE_IO=thread env var to force the thread pool backend on
any platform, enabling testing of thread pool code paths on Linux.
Revert process::exit — the hang is properly fixed now.

+++ src/io/backend/mod.rs
@@ -768,8 +768,14 @@ impl SyncBackend {
                 )
             }
         };
-        let fd =
-            unsafe { libc::openat(libc::AT_FDCWD, c_path.as_ptr(), flags, mode as libc::mode_t) };
+        let fd = unsafe {
+            libc::openat(
+                libc::AT_FDCWD,
+                c_path.as_ptr(),
+                flags,
+                libc::c_uint::from(mode as libc::mode_t),
+            )
+        };
         if fd < 0 {
             let err = std::io::Error::last_os_error();
             return (
